move TypeRegistries out fo API package
